Add basic setter/getter API

+++ gandhi-backend/src/main/kotlin/Data/Block.kt
@@ -2,4 +2,4 @@ package Data
 
 import java.util.Date
 
-data class Block(val number: Int, val timestamp: Date, val author: Address, val transactions: Map<Address, Transaction>)
+data class Block(val number: Long, val timestamp: Date, val author: Address, val transactions: Map<Address, Transaction>)
